// front/src/service/api/model/statusUpdateRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StatusUpdateRequest { 
    status: StatusUpdateRequest.StatusEnum;
}
export namespace StatusUpdateRequest {
    export const StatusEnum = {
        AContacter: 'A_CONTACTER',
        Contacte: 'CONTACTE',
        Entretien: 'ENTRETIEN',
        Accepte: 'ACCEPTE',
        Refuse: 'REFUSE'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}